CE-6605: fixed jenkins warnings

+++ src/codec/encoder/IMayaCallbacks.h
@@ -53,7 +53,7 @@ public:
 			const double* nrm, size_t nrmSize,
 			const uint32_t* faceSizes, size_t numFaces,
 			const uint32_t* indices, size_t indicesSize,
-			double const* const* uvs, size_t const* uvsSizes, uint32_t uvSetsCount,
+			double const* const* uvs, size_t const* uvsSizes, size_t uvSetsCount,
 			const uint32_t* faceRanges, size_t faceRangesSize,
 			const prt::AttributeMap** materials,
 			const prt::AttributeMap** reports,
